backend/requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.4,<5.0
djangorestframework>=3.14,<4.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
pillow>=10.4,<11.0
psycopg2-binary>=2.9,<3.0
